Add Rossler attractor

+++ software/o_c_REV/halvorsen_generator.h
@@ -77,12 +77,12 @@ class HalvorsenGenerator {
   }
 
  private:
-  int32_t x1_, y1_, z1_;
+  int64_t x1_, y1_, z1_;
   // double x1_, y1_, z1_;
-  int32_t rate1_;
-  int32_t x2_, y2_, z2_;
+  int64_t rate1_;
+  int64_t x2_, y2_, z2_;
   // double x2_, y2_, z2_;
-  int32_t rate2_;
+  int64_t rate2_;
 
   uint8_t out_c_, out_d_ ;
